+++ packages/webcodecs/src/convert-media.ts
@@ -13,16 +13,11 @@ import {getAudioEncoderConfig} from './audio-encoder-config';
 import type {ConvertMediaAudioCodec} from './codec-id';
 import {
 	codecNameToMatroskaAudioCodecId,
-	codecNameToMatroskaCodecId,
 	type ConvertMediaVideoCodec,
 } from './codec-id';
 import Error from './error-cause';
+import {makeVideoTrackHandler} from './on-video-track';
 import {resolveAudioAction} from './resolve-audio-action';
-import {resolveVideoAction} from './resolve-video-action';
-import {createVideoDecoder} from './video-decoder';
-import {getVideoDecoderConfigWithHardwareAcceleration} from './video-decoder-config';
-import {createVideoEncoder} from './video-encoder';
-import {getVideoEncoderConfig} from './video-encoder-config';
 import {withResolvers} from './with-resolvers';
 
 export type ConvertMediaState = {
@@ -104,134 +99,15 @@ export const convertMedia = async ({
 		canUseWebFs ? webFsWriter : bufferWriter,
 	);
 
-	const onVideoTrack: OnVideoTrack = async (track) => {
-		const videoEncoderConfig = await getVideoEncoderConfig({
-			codec: 'vp8',
-			height: track.displayAspectHeight,
-			width: track.displayAspectWidth,
-		});
-		const videoDecoderConfig =
-			await getVideoDecoderConfigWithHardwareAcceleration(track);
-		const videoOperation = await resolveVideoAction({
-			videoDecoderConfig,
-			videoEncoderConfig,
-		});
-
-		if (videoOperation === 'drop') {
-			return null;
-		}
-
-		if (videoOperation === 'copy') {
-			// TODO: Copy over video track
-
-			const videoTrack = await state.addTrack({
-				type: 'video',
-				// TODO: Copy over colors, CodecPrivate...
-				color: {
-					transferChracteristics: 'bt709',
-					matrixCoefficients: 'bt709',
-					primaries: 'bt709',
-					fullRange: true,
-				},
-				width: track.codedWidth,
-				height: track.codedHeight,
-				codecId: track.codec,
-			});
-			return (sample) => {
-				state.addSample(new EncodedVideoChunk(sample), videoTrack.trackNumber);
-				convertMediaState.decodedVideoFrames++;
-				onMediaStateUpdate?.({...convertMediaState});
-			};
-		}
-
-		if (videoEncoderConfig === null) {
-			abortConversion(
-				new Error(
-					`Could not configure video encoder of track ${track.trackId}`,
-				),
-			);
-			return null;
-		}
-
-		if (videoDecoderConfig === null) {
-			abortConversion(
-				new Error(
-					`Could not configure video decoder of track ${track.trackId}`,
-				),
-			);
-			return null;
-		}
-
-		const {trackNumber} = await state.addTrack({
-			type: 'video',
-			color: {
-				transferChracteristics: 'bt709',
-				matrixCoefficients: 'bt709',
-				primaries: 'bt709',
-				fullRange: true,
-			},
-			width: track.codedWidth,
-			height: track.codedHeight,
-			codecId: codecNameToMatroskaCodecId(videoCodec),
-		});
-
-		const videoEncoder = createVideoEncoder({
-			onChunk: async (chunk) => {
-				await state.addSample(chunk, trackNumber);
-				const newDuration = Math.round(
-					(chunk.timestamp + (chunk.duration ?? 0)) / 1000,
-				);
-				await state.updateDuration(newDuration);
-				convertMediaState.encodedVideoFrames++;
-				onMediaStateUpdate?.({...convertMediaState});
-			},
-			onError: (err) => {
-				abortConversion(
-					new Error(
-						`Video encoder of track ${track.trackId} failed (see .cause of this error)`,
-						{
-							cause: err,
-						},
-					),
-				);
-			},
-			signal: controller.signal,
-			config: videoEncoderConfig,
-		});
-
-		const videoDecoder = createVideoDecoder({
-			config: videoDecoderConfig,
-			onFrame: async (frame) => {
-				await onVideoFrame?.(frame, track);
-				await videoEncoder.encodeFrame(frame);
-				convertMediaState.decodedVideoFrames++;
-				onMediaStateUpdate?.({...convertMediaState});
-				frame.close();
-			},
-			onError: (err) => {
-				abortConversion(
-					new Error(
-						`Video decoder of track ${track.trackId} failed (see .cause of this error)`,
-						{
-							cause: err,
-						},
-					),
-				);
-			},
-			signal: controller.signal,
-		});
-
-		state.addWaitForFinishPromise(async () => {
-			await videoDecoder.waitForFinish();
-			await videoEncoder.waitForFinish();
-			videoDecoder.close();
-			videoEncoder.close();
-		});
-
-		return async (chunk) => {
-			await videoDecoder.processSample(chunk);
-		};
-	};
+	const onVideoTrack: OnVideoTrack = makeVideoTrackHandler({
+		state,
+		onVideoFrame: onVideoFrame ?? null,
+		onMediaStateUpdate: onMediaStateUpdate ?? null,
+		abortConversion,
+		convertMediaState,
+		controller,
+		videoCodec,
+	});
 
 	const onAudioTrack: OnAudioTrack = async (track) => {
 		const audioEncoderConfig = await getAudioEncoderConfig({
